LPS-58561 SF: Move tests for their own package
