+++ modules/portal/portal-portlet-container-test/test/integration/com/liferay/portal/portlet/container/upload/test/dependencies/file_upload.txt
